+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v1.0.5] - 2026-05-29
+
+### Added
+- `SWTools/TMP Font Asset Manager`에 TextMeshPro 폰트 에셋 성능 탭 추가
+- 성능 탭에서 아틀라스 메모리, 글리프, 문자, 폴백 체인, 동적 아틀라스, 머티리얼 프리셋 점검 지원
+- README에 TextMeshPro 폰트 에셋 성능 확인 가이드 추가
+
+### Changed
+- README 및 Changelog 버전 표기를 v1.0.5로 갱신
+
 ## [1.0.0] - 2026-03-19
 
 ### Added